[SCRUM-95] refactor: 서비스 계층 네이밍 컨벤션 통일
