+++ src/main/java/com/projectlyrics/server/domain/auth/api/AuthController.java
@@ -3,7 +3,7 @@ package com.projectlyrics.server.domain.auth.api;
 import com.projectlyrics.server.domain.auth.dto.request.UserLoginRequest;
 import com.projectlyrics.server.domain.common.dto.SuccessResponse;
 import com.projectlyrics.server.domain.user.dto.response.UserLoginResponse;
-import com.projectlyrics.server.domain.user.usecase.command.UserAuthUseCase;
+import com.projectlyrics.server.domain.user.service.UserCommandService;
 import com.projectlyrics.server.domain.common.message.SuccessMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class AuthController {
 
-  private final UserAuthUseCase userAuthUseCase;
+  private final UserCommandService userCommandService;
 
   @PostMapping("/login")
   public ResponseEntity<SuccessResponse<UserLoginResponse>> signIn(
@@ -30,7 +30,7 @@ public class AuthController {
         .status(HttpStatus.CREATED)
         .body(SuccessResponse.of(
             SuccessMessage.LOGIN_SUCCESS,
-            userAuthUseCase.signIn(socialAccessToken, loginRequest)
+            userCommandService.signIn(socialAccessToken, loginRequest)
         ));
   }
 }
