+++ src/main/java/com/projectlyrics/server/domain/user/service/UserCommandService.java
@@ -1,9 +1,9 @@
-package com.projectlyrics.server.domain.user.usecase.command;
+package com.projectlyrics.server.domain.user.service;
 
 import com.projectlyrics.server.domain.auth.dto.request.UserLoginRequest;
 import com.projectlyrics.server.domain.user.dto.response.UserLoginResponse;
 
-public interface UserAuthUseCase {
+public interface UserCommandService {
 
   UserLoginResponse signIn(String socialAccessToken, UserLoginRequest request);
 }
